+++ src/main/java/com/weave/weaveserver/repository/BelongRepository.java
@@ -3,5 +3,5 @@ package com.weave.weaveserver.repository;
 import com.weave.weaveserver.domain.Belong;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BelongRepository extends JpaRepository<Belong,Integer> {
+public interface BelongRepository extends JpaRepository<Belong, Long> {
 }
